+++ components/analytics-core/org.wso2.carbon.analytics.dataservice/src/main/java/org/wso2/carbon/analytics/dataservice/AnalyticsDataService.java
@@ -245,7 +245,7 @@ public interface AnalyticsDataService extends AnalyticsRecordReader {
      * @return the count of the records which match the drilldown query
      * @throws AnalyticsIndexException
      */
-    public int drillDownSearchCount(int tenantId, AnalyticsDrillDownRequest drillDownRequest)
+    public double drillDownSearchCount(int tenantId, AnalyticsDrillDownRequest drillDownRequest)
             throws AnalyticsIndexException;
 
     /**
