+++ components/analytics-data-connectors/org.wso2.carbon.analytics.datasource.rdbms/src/test/java/org/wso2/carbon/analytics/datasource/rdbms/AnalyticsDataServiceTest.java
@@ -566,7 +566,7 @@ public class AnalyticsDataServiceTest implements GroupEventListener {
 
         System.out.println("* Search Result Count: " + results.size() + " Time: " + (end - start) + " ms.");
         start = System.currentTimeMillis();
-        int count = 0;
+        double count = 0;
         for (String tableName : tableNames) {
             drillDownRequest.setTableName(tableName);
             drillDownRequest.setRecordStartIndex(0);
@@ -713,9 +713,9 @@ public class AnalyticsDataServiceTest implements GroupEventListener {
         Assert.assertEquals(results.size(), 75);
         System.out.println("* Drilldown Result Count: " + results.size() + " Time: " + (end - start) + " ms.");
         start = System.currentTimeMillis();
-        int count = this.service.drillDownSearchCount(tenantId, drillDownRequest);
+        double count = this.service.drillDownSearchCount(tenantId, drillDownRequest);
         end = System.currentTimeMillis();
-        Assert.assertEquals(count, 50000);
+        Assert.assertEquals(count, 50000.0);
         System.out.println("* DrilldownSearch Result Count: " + count + " Time: " + (end - start) + " ms.");
         this.cleanupTable(tenantId, tableName);
         System.out.println("\n************** END ANALYTICS DS (WITH FACET INDEXING - MULTIPLE THREADS, H2-FILE) PERF TEST **************");
@@ -774,7 +774,7 @@ public class AnalyticsDataServiceTest implements GroupEventListener {
         Assert.assertTrue(results.size() <= 75 * tableNames.length);
         System.out.println("* Drilldown Result Count: " + results.size() + " Time: " + (end - start) + " ms.");
         start = System.currentTimeMillis();
-        int count = 0;
+        double count = 0;
         for (String tableName : tableNames) {
             AnalyticsDrillDownRequest drillDownRequest = new AnalyticsDrillDownRequest();
             drillDownRequest.setTableName(tableName);
